fixlib/pyproject.toml
[project]
name = "fixinventorylib"
version = "4.0.0"
authors = [{ name = "Some Engineering Inc." }]
description = "Resoto common library."
license = { text = "AGPLv3" }
requires-python = ">=3.9"
classifiers = [
    # Current project status
    "Development Status :: 5 - Production/Stable",
    # Audience
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    # License information
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    # Supported python versions
    "Programming Language :: Python :: 3.9",
    # Supported OS's
    "Operating System :: POSIX :: Linux",
    "Operating System :: Unix",
    # Extra metadata
    "Environment :: Console",
    "Natural Language :: English",
    "Topic :: Security",
    "Topic :: Utilities",
]
readme = { file = "README.md", content-type = "text/markdown" }
keywords = ["cloud security"]

dependencies = [
    "CherryPy",
    "Pint",
    "PyJWT",
    "PyYAML",
    "aiohttp[speedups]",
    "attrs",
    "cattrs",
    "cryptography",
    "frozendict",
    "isodate",
    "jsons",
    "networkx",
    "parsy",
    "prometheus-client",
    "psutil",
    "python-dateutil",
    "requests",
    "typeguard",
    "tzdata",
    "tzlocal",
    "websocket-client"
]

[project.optional-dependencies]
test = [
    "black",
    "coverage",
    "flake8",
    "hypothesis",
    "mypy",
    "pep8-naming",
    "pylint",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-runner",
    "tox",
    "types-PyYAML",
    "types-python-dateutil",
    "types-pytz",
    "types-requests",
    "types-setuptools",
    "types-tzlocal",
    "wheel",
]

[project.urls]
Documentation = "https://inventory.fix.security"
Source = "https://github.com/someengineering/fixinventory/tree/main/fixlib"

[tool.setuptools.package-data]
fixlib = ["py.typed"]

[build-system]
requires = ["setuptools>=67.8.0", "wheel>=0.40.0", "build>=0.10.0"]
build-backend = "setuptools.build_meta"

